requirements.txt
```text
kubernetes
jenkins-python
prometheus_client
rich
pyyaml
ruamel.yaml
python-nmap
requests
dataclasses
```